LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.strokeRect.clip.worker-expected.txt


PASS strokeRect is affected by clipping regions 
